5.1.5

See merge request BI/magreport-internal-2.1!81

+++ magreport-backend/src/main/java/ru/magnit/magreportbackend/metrics_function/impl/MinIntegerFunction.java
@@ -4,17 +4,17 @@ import ru.magnit.magreportbackend.metrics_function.MetricsFunction;
 
 public class MinIntegerFunction implements MetricsFunction {
 
-    private Double minValue;
+    private Long minValue;
 
     @Override
     public void addValue(String value, int col, int row) {
         if (value != null)
-            minValue = minValue== null ? Double.parseDouble(value) : Double.min(minValue, Double.parseDouble(value));
+            minValue = minValue== null ? Long.parseLong(value) : Long.min(minValue, Long.parseLong(value));
 
     }
 
     @Override
     public String calculate(int col, int row) {
-        return minValue == null ? "" : String.valueOf(minValue.intValue());
+        return minValue == null ? "" : String.valueOf(minValue.longValue());
     }
 }
